Rename class files from CamelCase.py to snake_case.py
