+++ src/wtfjson/__init__.py
@@ -6,6 +6,6 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from .DictInput import DictInput
-from .ListInput import ListInput
+from .dict_input import DictInput
+from .list_input import ListInput
