+++ src/wtfjson/dict_input.py
