+++ src/wtfjson/fields/__init__.py
@@ -6,15 +6,15 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from .Field import Field
-from .UnboundField import UnboundField
-from .BooleanField import BooleanField
-from .StringField import StringField
-from .IntegerField import IntegerField
-from .EnumField import EnumField
-from .DecimalField import DecimalField
-from .ListField import ListField
-from .ObjectField import ObjectField
-from .DateField import DateField
-from .DateTimeField import DateTimeField
-from .FloatField import FloatField
+from .field import Field
+from .unbound_field import UnboundField
+from .boolean_field import BooleanField
+from .string_field import StringField
+from .integer_field import IntegerField
+from .enum_field import EnumField
+from .decimal_field import DecimalField
+from .list_field import ListField
+from .object_field import ObjectField
+from .date_field import DateField
+from .date_time_field import DateTimeField
+from .float_field import FloatField
