+++ src/wtfjson/fields/date_field.py
@@ -7,20 +7,23 @@ Use of this source code is governed by an MIT-style license that can be found in
 """
 
 from typing import Union
+from datetime import date
+
 from ..fields import Field
-from ..validators import Type
+from ..validators import Type, Date
 from ..util import UnsetValue
 
 
-class FloatField(Field):
+class DateField(Field):
     pre_validators = [
-        Type(data_type=float)
+        Type(data_type=str),
+        Date()
     ]
 
     @property
-    def data(self) -> Union[float, UnsetValue]:
+    def data(self) -> Union[date, UnsetValue]:
         return super().data
 
     @property
-    def out(self) -> Union[float, UnsetValue]:
+    def out(self) -> Union[date, UnsetValue]:
         return super().out
