+++ src/wtfjson/fields/date_time_field.py
@@ -6,27 +6,26 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from enum import Enum
 from typing import Union
+from datetime import datetime
+
 from ..fields import Field
-from ..validators import Type, EnumValidator
+from ..validators import Type, DateTime
 from ..util import UnsetValue
 
 
-class EnumField(Field):
-
-    def __init__(self, enum, *args, **kwargs):
+class DateTimeField(Field):
+    def __init__(self, localized: bool = False, accept_utc=False, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.enum = enum
-        self.default_validators = [
+        self.pre_validators = [
             Type(data_type=str),
-            EnumValidator(enum=enum)
+            DateTime(localized=localized, accept_utc=accept_utc)
         ]
 
     @property
-    def data(self) -> Union[Enum, UnsetValue]:
+    def data(self) -> Union[datetime, UnsetValue]:
         return super().data
 
     @property
-    def out(self) -> Union[Enum, UnsetValue]:
+    def out(self) -> Union[datetime, UnsetValue]:
         return super().out
