+++ src/wtfjson/fields/field.py
@@ -9,12 +9,12 @@ Use of this source code is governed by an MIT-style license that can be found in
 from abc import ABC
 from enum import Enum
 from copy import deepcopy
-from typing import List, Callable, Optional, Any
+from typing import List, Callable, Optional, Any, TYPE_CHECKING
+
 from ..util import unset_value
 from ..exceptions import ValidationError, StopValidation, ClearValidation
-from .UnboundField import UnboundField
+from .unbound_field import UnboundField
 
-from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     import Form
