+++ src/wtfjson/fields/string_field.py
@@ -7,6 +7,7 @@ Use of this source code is governed by an MIT-style license that can be found in
 """
 
 from typing import Union
+
 from ..fields import Field
 from ..validators import Type
 from ..util import UnsetValue
