+++ src/wtfjson/validators/__init__.py
@@ -6,19 +6,19 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from .Validator import Validator
-from .Type import Type
-from .EnumValidator import EnumValidator
-from .DecimalValidator import DecimalValidator
-from .ListLength import ListLength
-from .Email import Email
-from .Regexp import Regexp
-from .URL import URL
-from .NoneOf import NoneOf
-from .AnyOf import AnyOf
-from .Length import Length
-from .NumberRange import NumberRange
-from .InputRequired import InputRequired
-from .Date import Date
-from .DateTime import DateTime
-from .DateTimeRange import DateTimeRange
+from .validator import Validator
+from .type import Type
+from .enum_validator import EnumValidator
+from .decimal_validator import DecimalValidator
+from .list_length import ListLength
+from .email import Email
+from .regexp import Regexp
+from .url import URL
+from .none_of import NoneOf
+from .any_of import AnyOf
+from .length import Length
+from .number_range import NumberRange
+from .input_required import InputRequired
+from .date import Date
+from .date_time import DateTime
+from .date_time_range import DateTimeRange
