+++ src/wtfjson/validators/any_of.py
@@ -6,22 +6,24 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from typing import Any, Union, TYPE_CHECKING
-from email_validator import validate_email, EmailNotValidError
+from typing import Any, Optional, TYPE_CHECKING, Union
 
+from ..fields import Field
 from ..validators import Validator
 from ..exceptions import ValidationError
-from ..fields import Field
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
-class Email(Validator):
-    default_message = 'invalid email'
+class AnyOf(Validator):
+    default_message = 'is not in any-of list'  # TODO: better message
+
+    def __init__(self, any_of: list, message: Optional[str] = None):
+        super().__init__(message)
+        self.any_of = any_of
 
     def __call__(self, value: Any, form: Union['DictInput', 'ListInput'], field: Field):
-        try:
-            validate_email(value)
-        except EmailNotValidError:
+        if value not in self.any_of:
             raise ValidationError(self.message)
