+++ src/wtfjson/validators/decimal_validator.py
@@ -12,9 +12,10 @@ from decimal import Decimal, InvalidOperation
 from ..fields import Field
 from ..validators import Validator
 from ..exceptions import ValidationError
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
 class DecimalValidator(Validator):
