+++ src/wtfjson/validators/input_required.py
@@ -7,6 +7,7 @@ Use of this source code is governed by an MIT-style license that can be found in
 """
 
 from typing import Optional
+
 from ..validators import Length
