+++ src/wtfjson/validators/list_length.py
@@ -11,9 +11,10 @@ from typing import Optional, Any, Union, TYPE_CHECKING
 from ..fields import Field
 from ..validators import Validator
 from ..exceptions import ValidationError
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
 class ListLength(Validator):
