+++ src/wtfjson/validators/none_of.py
@@ -6,25 +6,24 @@ Copyright (c) 2021, binary butterfly GmbH
 Use of this source code is governed by an MIT-style license that can be found in the LICENSE.txt.
 """
 
-from enum import Enum
-from typing import Optional, Any, Union, TYPE_CHECKING
+from typing import Any, Optional, Union, TYPE_CHECKING
 
 from ..fields import Field
 from ..validators import Validator
 from ..exceptions import ValidationError
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
-class EnumValidator(Validator):
-    default_message = 'value not in enum'
+class NoneOf(Validator):
+    default_message = 'is in none-of list'  # TODO: better message
 
-    def __init__(self, enum: Enum, message: Optional[str] = None):
+    def __init__(self, none_of: list, message: Optional[str] = None):
         super().__init__(message)
-        self.enum = enum
+        self.none_of = none_of
 
     def __call__(self, value: Any, form: Union['DictInput', 'ListInput'], field: Field):
-        if type(field.data_processed) is not str or not hasattr(self.enum, field.data_processed):
+        if value in self.none_of:
             raise ValidationError(self.message)
-        field.data_processed = getattr(self.enum, field.data_processed)
