+++ src/wtfjson/validators/number_range.py
@@ -11,9 +11,10 @@ from typing import Any, Optional, Union, TYPE_CHECKING
 from ..fields import Field
 from ..validators import Validator
 from ..exceptions import ValidationError, InvalidData
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
 class NumberRange(Validator):
