+++ src/wtfjson/validators/type.py
@@ -11,9 +11,10 @@ from typing import Optional, Any, Union, TYPE_CHECKING
 from ..fields import Field
 from ..validators import Validator
 from ..exceptions import StopValidation
+
 if TYPE_CHECKING:
-    from ..DictInput import DictInput
-    from ..ListInput import ListInput
+    from ..dict_input import DictInput
+    from ..list_input import ListInput
 
 
 class Type(Validator):
